+++ llama_index/evaluation/relevancy.py
@@ -1,6 +1,7 @@
 """Relevancy evaluation."""
 from __future__ import annotations
 
+import asyncio
 from typing import Any, Sequence
 
 from llama_index import ServiceContext
@@ -96,6 +97,7 @@ class RelevancyEvaluator(BaseEvaluator):
         query: str | None = None,
         response: str | None = None,
         contexts: Sequence[str] | None = None,
+        sleep_time_in_seconds: int = 0,
         **kwargs: Any,
     ) -> EvaluationResult:
         """Evaluate whether the contexts and response are relevant to the query."""
@@ -108,6 +110,8 @@ class RelevancyEvaluator(BaseEvaluator):
         index = SummaryIndex.from_documents(docs, service_context=self._service_context)
 
         query_response = f"Question: {query}\nResponse: {response}"
+
+        await asyncio.sleep(sleep_time_in_seconds)
 
         query_engine = index.as_query_engine(
             text_qa_template=self._eval_template,