+++ llama_index/indices/vector_store/base.py
@@ -292,7 +292,7 @@ class VectorStoreIndex(BaseIndex[IndexDict]):
         self, ref_doc_id: str, delete_from_docstore: bool = False, **delete_kwargs: Any
     ) -> None:
         """Delete a document and it's nodes by using ref_doc_id."""
-        self._vector_store.delete(ref_doc_id)
+        self._vector_store.delete(ref_doc_id, **delete_kwargs)
 
         # delete from index_struct only if needed
         if not self._vector_store.stores_text or self._store_nodes_override: