+++ llama_index/multi_modal_llms/replicate_multi_modal.py
@@ -1,3 +1,4 @@
+import logging
 from typing import Any, Callable, Dict, Optional, Sequence
 
 from llama_index.bridge.pydantic import Field, PrivateAttr
@@ -21,8 +22,10 @@ from llama_index.multi_modal_llms import (
 )
 from llama_index.schema import ImageDocument
 
+_logger = logging.getLogger(__name__)
+
 REPLICATE_MULTI_MODAL_LLM_MODELS = {
-    "llava-13b": "yorickvp/llava-13b:2facb4a474a0462c15041b78b1ad70952ea46b5ec6ad29583c0b29dbd4249591",
+    "llava-13b": "yorickvp/llava-13b:e272157381e2a3bf12df3a8edd1f38d1dbd736bbb7437277c8b34175f8fce358",
     "fuyu-8b": "lucataco/fuyu-8b:42f23bc876570a46f5a90737086fbc4c3f79dd11753a28eaa39544dd391815e9",
     "minigpt-4": "daanelson/minigpt-4:b96a2f33cc8e4b0aa23eacfce731b9c41a7d9466d9ed4e167375587b54db9423",
     "cogvlm": "naklecha/cogvlm:ec3886f9ea85dd0aee216585be5e6d07b04c9650f7b8b08363a14eb89e207eb2",
@@ -170,8 +173,9 @@ class ReplicateMultiModal(MultiModalLLM):
 
         # TODO: at the current moment, only support uploading one image document
         if len(image_documents) > 1:
-            raise NotImplementedError(
+            _logger.warning(
                 "ReplicateMultiModal currently only supports uploading one image document"
+                "we are using the first image document for completion."
             )
 
         prompt = self._completion_to_prompt(prompt)
@@ -238,8 +242,9 @@ class ReplicateMultiModal(MultiModalLLM):
 
         # TODO: at the current moment, only support uploading one image document
         if len(image_documents) > 1:
-            raise NotImplementedError(
+            _logger.warning(
                 "ReplicateMultiModal currently only supports uploading one image document"
+                "we are using the first image document for completion."
             )
 
         prompt = self._completion_to_prompt(prompt)